ERT-747: Changed link_segments default value
Faults are often not singly connected, when adding a barrier in a fault
block due to faults it is possible to link the the different parts of a
fault with a barrier, that is controlled by the optional keyword
link_segments.

With this commit the default has changed to False, that is because the
fault blocks often do not follow the natural link extension between
fault parts, and then things blow up.