ERT-868: Changed local_obsset -> local_obsdata
With this commit the old local_obsset class is removed, and everything
is implemented with the local_obsdata type. There is still some
unresolved conflict here with respect to where the time steps active for
observations are configured:

 1: The local_obsdata_node instances carry information of which time
    steps are active for this particular observation.

 2: The enkf_main_UPDATE() routine takes the time steps used for update
    as input arguments; this is inheritance from EnKF.

Ideally the enkf_main_UPDATE() routine should *not* take direct
timestep input; that still remains.