fix the failing python unit tests
this is done by fixing the findDevRoot() method of the
SourceEnumerator python class: before, this method just went up the
directory which contained the 'source_enumerator.py' file and stopped
at the first directory which had a 'devel' member. this approach
obviously fails if 'devel' is replaced by '.' because every directory
has a member called '.' (i.e., itself; note that the old approach also
fails if a file called 'devel' is added to e.g. the build directory or
the build directory is not a sibling of the `devel` directory). the
quick fix is to stop if the directory contains the file called
"libecl/include/ert/ecl/ecl_region.h", but I think this approach still
is pretty much a kludge.